requirements.txt
openai
pdfplumber
pandas
pinecone-client
streamlit
